requirements.txt
fastapi
farm-haystack
uvicorn
torch
torchvision
torchaudio